backend/requirements.txt
# FastAPI framework and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for settings and validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Telegram bot
python-telegram-bot[webhooks]==20.7

# OpenAI/LLM
openai==1.6.1

# HTTP client
httpx==0.25.2

# JWT authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Stripe payments
stripe==8.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==24.1.1
isort==5.13.2
ruff==0.1.14
mypy==1.8.0

# Additional type stubs
types-redis==4.6.0.20240106
